+++ Changelog.md
@@ -6,6 +6,8 @@ Additions:
 
 Changes:
 
+- Add active toggle for timeseries that determines if they should be updated.
+
 Fixes:
 
 ## 0.3.1 - 03/08/2021